pyproject.toml
[tool.poetry]
name = "pxf-framework-core"
version = "0.1.0"
description = ""
authors = ["NewWars"]
readme = "README.md"
packages = [
    {include = "pxf_framework_core"},
    {include = "pxf_framework_core/py.typed"}
]


[tool.poetry.dependencies]
python = "^3.11"
dataclasses-json = "*"
hjson="*"
events = "*"
python-i18n = "*"
memory-profiler = "*"


[tool.poetry.dev-dependencies]
pre-commit = "*"

[tool.mypy]
show_error_codes = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
